+++ src/Angor/Avalonia/AngorApp/Controls/IconButton.axaml
@@ -1,20 +1,21 @@
 <ResourceDictionary xmlns="https://github.com/avaloniaui"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:controls="using:AngorApp.Views"
-                    xmlns:avalonia="https://github.com/projektanker/icons.avalonia">
+                    xmlns:avalonia="https://github.com/projektanker/icons.avalonia"
+                    xmlns:controls1="clr-namespace:AngorApp.Controls">
 
     <Design.PreviewWith>
-        <controls:IconButton Content="THIS IS A BUTTON">
-            <controls:IconButton.Icon>
+        <controls1:IconButton Content="THIS IS A BUTTON">
+            <controls1:IconButton.Icon>
                 <avalonia:Icon Value="fa-wallet" />
-            </controls:IconButton.Icon>
-        </controls:IconButton>
+            </controls1:IconButton.Icon>
+        </controls1:IconButton>
     </Design.PreviewWith>
 
-    <ControlTheme TargetType="controls:IconButton" x:Key="{x:Type controls:IconButton}" BasedOn="{StaticResource {x:Type Button}}">
+    <ControlTheme TargetType="controls1:IconButton" x:Key="{x:Type controls1:IconButton}" BasedOn="{StaticResource {x:Type Button}}">
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="controls:IconButton">
+                <ControlTemplate TargetType="controls1:IconButton">
                     <Button
                         Padding="{TemplateBinding Padding}"
                         Margin="{TemplateBinding Margin}"
