+++ src/Angor/Avalonia/AngorApp/Controls/IconButton.axaml.cs
@@ -1,7 +1,6 @@
 using Avalonia;
-using Avalonia.Controls;
 
-namespace AngorApp.Views;
+namespace AngorApp.Controls;
 
 public class IconButton : Button
 {
