+++ src/Angor/Avalonia/AngorApp/Controls/Pane.axaml
@@ -7,13 +7,14 @@
     
     <ControlTheme x:Key="{x:Type c:Pane}" TargetType="c:Pane">
         <Setter Property="ClipToBounds" Value="False" />
+        <Setter Property="Padding" Value="10" />
         <Setter Property="Template">
             <ControlTemplate>
                 <Border Background="{StaticResource PaneBackground}" CornerRadius="10" ClipToBounds="True">
                     <Border.Effect>
                         <DropShadowEffect Color="Black" BlurRadius="10" Opacity="0.5" OffsetX="2" OffsetY="2" />
                     </Border.Effect>
-                    <ContentPresenter Margin="10" Content="{TemplateBinding Content}" />
+                    <ContentPresenter Margin="{TemplateBinding Padding}" Content="{TemplateBinding Content}" />
                 </Border>
             </ControlTemplate>
         </Setter>
