+++ src/Angor/Avalonia/AngorApp/Controls/Pane.axaml.cs
@@ -1,5 +1,3 @@
-using Avalonia.Controls;
-
 namespace AngorApp.Controls;
 
 public class Pane : ContentControl;
