+++ src/Angor/Avalonia/AngorApp/GlobalUsings.cs
@@ -1,0 +1,7 @@
+// Global using directives
+
+global using System;
+global using System.Collections.Generic;
+global using System.Reactive;
+global using Avalonia.Controls;
+global using ReactiveUI;
