+++ src/Angor/Avalonia/AngorApp/Sections/Browse/BrowseView.axaml.cs
@@ -1,5 +1,3 @@
-using Avalonia.Controls;
-
 namespace AngorApp.Sections.Browse;
 
 public partial class BrowseView : UserControl
