+++ src/Angor/Avalonia/AngorApp/Sections/Browse/BrowseViewModel.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Reactive;
-using ReactiveUI;
+using AngorApp.Services;
 
 namespace AngorApp.Sections.Browse;
