+++ src/Angor/Avalonia/AngorApp/Sections/Founder/FounderView.axaml.cs
@@ -1,5 +1,3 @@
-using Avalonia.Controls;
-
 namespace AngorApp.Sections.Founder;
 
 public partial class FounderView : UserControl
