+++ src/Angor/Avalonia/AngorApp/Sections/Home/HomeView.axaml.cs
@@ -1,5 +1,4 @@
 using Avalonia;
-using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
 namespace AngorApp.Sections.Home;
