+++ src/Angor/Avalonia/AngorApp/Sections/Home/HomeViewModel.cs
@@ -1,5 +1,3 @@
-using ReactiveUI;
-
 namespace AngorApp.Sections.Home;
 
 public class HomeViewModel : ReactiveObject
