+++ src/Angor/Avalonia/AngorApp/Sections/Portfolio/IPortfolioViewModel.cs
@@ -1,0 +1,6 @@
+namespace AngorApp.Sections.Portfolio;
+
+public interface IPortfolioViewModel
+{
+    IReadOnlyCollection<PortfolioItem> Items { get; }
+}
