+++ src/Angor/Avalonia/AngorApp/Sections/Portfolio/PorfolioView.axaml
@@ -1,0 +1,91 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:controls="clr-namespace:AngorApp.Controls"
+             xmlns:i="https://github.com/projektanker/icons.avalonia"
+             mc:Ignorable="d" d:DesignWidth="800"
+             x:Class="AngorApp.Sections.Portfolio.PortfolioView"
+             xmlns:views="clr-namespace:AngorApp.Views"
+             xmlns:founder="clr-namespace:AngorApp.Sections.Portfolio"
+             x:DataType="founder:IPortfolioViewModel" ClipToBounds="False">
+
+    <Design.DataContext>
+        <founder:PortfolioViewModelDesign />
+    </Design.DataContext>
+
+    <UserControl.Styles>
+
+        <Style Selector="TextBox">
+            <Setter Property="BorderThickness" Value="1" />
+            <Setter Property="BorderBrush" Value="{StaticResource Level3}" />
+            <Setter Property="VerticalContentAlignment" Value="Center" />
+        </Style>
+
+        <Style Selector="i|Icon">
+            <Setter Property="FontSize" Value="25" />
+        </Style>
+        <Style Selector="Button">
+            <Setter Property="CornerRadius" Value="10" />
+            <Setter Property="Padding" Value="10" />
+            <Setter Property="HorizontalContentAlignment" Value="Center" />
+            <Setter Property="VerticalContentAlignment" Value="Center" />
+            <Setter Property="TextBlock.FontWeight" Value="Bold" />
+            <Setter Property="Foreground" Value="{StaticResource Level5}" />
+        </Style>
+        <Style Selector="TextBlock">
+            <Setter Property="VerticalAlignment" Value="Center" />
+        </Style>
+        
+        <Style Selector="ItemsControl#PortfolioItems TextBlock">
+            <Setter Property="HorizontalAlignment" Value="Center" />
+        </Style>
+        <Style Selector="ItemsControl#PortfolioItems controls|Pane">
+            <Setter Property="Margin" Value="20" />
+        </Style>
+        <Style Selector="ItemsControl#PortfolioItems controls|Pane > StackPanel">
+            <Setter Property="Spacing" Value="10" />
+            <Setter Property="Margin" Value="20" />
+        </Style>
+        
+    </UserControl.Styles>
+
+    <StackPanel Spacing="20">
+
+        <controls:Pane Padding="15" CornerRadius="5">
+            <DockPanel>
+                <i:Icon Value="fa-solid fa-hand-holding-dollar" Margin="10" />
+                <controls:IconButton DockPanel.Dock="Right" VerticalAlignment="Top" Content="PENALTIES">
+                    <controls:IconButton.Icon>
+                        <i:Icon Value="fa-regular fa-face-frown" FontSize="20" />
+                    </controls:IconButton.Icon>
+                </controls:IconButton>
+                <StackPanel Spacing="4" DockPanel.Dock="Left">
+                    <TextBlock Classes="PaneTitle" Text="Portfolio" FontWeight="Regular" FontSize="18" />
+                    <TextBlock TextWrapping="Wrap" Classes="PaneTitle" Text="To create a new project or view your existing projects, an on-chain transaction and a Nostr DID are required" FontWeight="Regular" VerticalAlignment="Center"
+                               FontSize="14" />
+                </StackPanel>
+            </DockPanel>
+        </controls:Pane>
+
+        <ItemsControl x:Name="PortfolioItems" ItemsSource="{Binding Items}" Margin="-20">
+            <ItemsControl.ItemTemplate>
+                <DataTemplate>
+                    <controls:Pane>
+                        <StackPanel>
+                            <i:Icon Value="fa-solid fa-money-bills"></i:Icon>
+                            <TextBlock Classes="PaneTitle" Text="Funded projects"></TextBlock>
+                            <TextBlock Classes="Regular" Text="0" />
+                        </StackPanel>
+                    </controls:Pane>
+                </DataTemplate>
+            </ItemsControl.ItemTemplate>
+            <ItemsControl.ItemsPanel>
+                <ItemsPanelTemplate>
+                    <UniformGrid />
+                </ItemsPanelTemplate>
+            </ItemsControl.ItemsPanel>
+        </ItemsControl>
+
+    </StackPanel>
+</UserControl>
