+++ src/Angor/Avalonia/AngorApp/Sections/Portfolio/PortfolioItem.cs
@@ -1,0 +1,3 @@
+namespace AngorApp.Sections.Portfolio;
+
+public record PortfolioItem(string Title, string Subtitle);
