+++ src/Angor/Avalonia/AngorApp/Sections/Portfolio/PortfolioView.cs
@@ -1,0 +1,9 @@
+namespace AngorApp.Sections.Portfolio;
+
+public partial class PortfolioView : UserControl
+{
+    public PortfolioView()
+    {
+        InitializeComponent();
+    }
+}
