+++ src/Angor/Avalonia/AngorApp/Sections/Portfolio/PortfolioViewModel.cs
@@ -1,0 +1,17 @@
+namespace AngorApp.Sections.Portfolio;
+
+public class PortfolioViewModel : ReactiveObject, IPortfolioViewModel
+{
+    public PortfolioViewModel()
+    {
+        Items =
+        [
+            new PortfolioItem("Ariton", "0"),
+            new PortfolioItem("Total invested", "0 TBTC"),
+            new PortfolioItem("Wallet", "0 TBTC"),
+            new PortfolioItem("In Recovery", "0 TBTC"),
+        ];
+    }
+
+    public IReadOnlyCollection<PortfolioItem> Items { get; }
+}
