+++ src/Angor/Avalonia/AngorApp/Sections/Shell/MainView.axaml
@@ -3,8 +3,8 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:shell="clr-namespace:AngorApp.Sections.Shell"
-             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="500"
-             x:Class="AngorApp.Views.MainView"
+             mc:Ignorable="d" d:DesignWidth="1000" d:DesignHeight="700"
+             x:Class="AngorApp.Sections.Shell.MainView"
              x:DataType="shell:MainViewModel">
 
     <Design.DataContext>
