+++ src/Angor/Avalonia/AngorApp/Sections/Shell/MainWindow.axaml
@@ -3,11 +3,12 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:views="clr-namespace:AngorApp.Views"
+        xmlns:shell="clr-namespace:AngorApp.Sections.Shell"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
         Width="800"
         Height="600"
-        x:Class="AngorApp.Views.MainWindow"
+        x:Class="AngorApp.Sections.Shell.MainWindow"
         Icon="/Assets/avalonia-logo.ico"
         Title="AngorApp">
-        <views:MainView />
+        <shell:MainView />
 </Window>
