+++ src/Angor/Avalonia/AngorApp/Sections/Shell/MainWindow.axaml.cs
@@ -1,6 +1,4 @@
-using Avalonia.Controls;
-
-namespace AngorApp.Views;
+namespace AngorApp.Sections.Shell;
 
 public partial class MainWindow : Window
 {
