+++ src/Angor/Avalonia/AngorApp/Sections/Shell/Separator.cs
@@ -1,0 +1,3 @@
+namespace AngorApp.Sections.Shell;
+
+public class Separator : SectionBase;
