+++ src/Angor/Avalonia/AngorApp/Sections/Shell/Sidebar.axaml.cs
@@ -1,5 +1,3 @@
-using Avalonia.Controls;
-
 namespace AngorApp.Sections.Shell;
 
 public partial class Sidebar : UserControl
