+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/WalletView.axaml.cs
@@ -1,5 +1,3 @@
-using Avalonia.Controls;
-
 namespace AngorApp.Sections.Wallet;
 
 public partial class WalletView : UserControl
