+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/WalletViewModel.cs
@@ -1,5 +1,3 @@
-using ReactiveUI;
-
 namespace AngorApp.Sections.Wallet;
 
 public class WalletViewModel : ReactiveObject
