+++ src/Angor/Avalonia/AngorApp/Services/ILauncherService.cs
@@ -1,0 +1,8 @@
+using System.Threading.Tasks;
+
+namespace AngorApp.Services;
+
+public interface ILauncherService
+{
+    Task Launch(Uri uri);
+}
