+++ src/Angor/Avalonia/AngorApp/Services/NoopLauncherService.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+
+namespace AngorApp.Services;
+
+public class NoopLauncherService : ILauncherService
+{
+    public Task Launch(Uri uri)
+    {
+        return Task.CompletedTask;
+    }
+}
