+++ src/Angor/Avalonia/AngorApp/Services/UIServices.cs
@@ -1,0 +1,11 @@
+namespace AngorApp.Services;
+
+public class UIServices
+{
+    public ILauncherService LauncherService { get; }
+
+    public UIServices(ILauncherService launcherService)
+    {
+        LauncherService = launcherService;
+    }
+}
